Split HBase server target construction
